Add files via upload
What is customer segmentation?

It is the process of dividing customers into various groups based on various factors that include demographics, salary, age etc. This is done so that the marketing team can plan the marketing strategy accordingly for each group

What is Unsupervised Learning?

Unsupervised learning is the training of a machine using information that is neither classified nor labeled and allowing the algorithm to act on that information without guidance. Here the task of the machine is to group unsorted information according to similarities, patterns, and differences without any prior training of data.
Unlike supervised learning, no teacher is provided that means no training will be given to the machine. Therefore the machine is restricted to find the hidden structure in unlabeled data by itself.

Problem Statement

Understand the target customers for the marketing team to plan strategy

Context

a) To identify the most important shopping groups based on
Income
Age
Shopping Score
b) Identify the ideal number of groups with a label for each

Approach
1) Perform EDA
2) Use KMeans Cluttering Algorithm to create segments
3) Use Summary Statistics on the clusters